LayoutTests/platform/chromium-linux/picture/working/picture-display-image-unset-allows-cached-image-load-expected.txt

CONSOLE MESSAGE: Invalid name for preference: WebKitUsesPageCachePreferenceKey
Test that WebKitDisplayImagesKey unset does not block image in cache.
PASS
